pyproject.toml
[tool.poetry]
name = "compiler-domjudge"
version = "0.1.0"
description = ""
authors = ["Yuki123456788 <yuki11835@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
